fetch all data opration using Rowmapper

+++ src/main/java/com/spring/jdbc/App.java
@@ -6,6 +6,7 @@ import java.util.Scanner;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.support.ClassPathXmlApplicationContext;
 
+import com.mysql.cj.protocol.x.SyncFlushDeflaterOutputStream;
 import com.spring.jdbc.dao.StudentDao;
 import com.spring.jdbc.entites.Student;
 
@@ -15,8 +16,7 @@ public class App {
 
 		System.out.println("My program started .....");
 
-		ApplicationContext context =
-				new ClassPathXmlApplicationContext("com/spring/jdbc/config.xml");
+		ApplicationContext context = new ClassPathXmlApplicationContext("com/spring/jdbc/config.xml");
 
 		StudentDao dao = context.getBean("studentDao", StudentDao.class);
 
@@ -28,8 +28,9 @@ public class App {
 			System.out.println("1. Insert");
 			System.out.println("2. Update");
 			System.out.println("3. Delete");
-			System.out.println("4. Select All");
-			System.out.println("5. Exit");
+			System.out.println("4. Select One");
+			System.out.println("5. Select All");
+			System.out.println("6. Exit");
 
 			System.out.println("Enter your choice:");
 			int choice = sc.nextInt();
@@ -46,22 +47,22 @@ public class App {
 
 				System.out.println("Enter City:");
 				String city1 = sc.nextLine();
-				
+
 				if (id1 <= 0) {
-				    System.out.println("Invalid Id. Must be greater than 0.");
-				    break;
+					System.out.println("Invalid Id. Must be greater than 0.");
+					break;
 				}
 
 				if (name1.trim().isEmpty()) {
-				    System.out.println("Name cannot be empty.");
-				    break;
+					System.out.println("Name cannot be empty.");
+					break;
 				}
 
 				if (city1.trim().isEmpty()) {
-				    System.out.println("City cannot be empty.");
-				    break;
+					System.out.println("City cannot be empty.");
+					break;
 				}
-				
+
 				Student student1 = new Student(id1, name1, city1);
 				int insertResult = dao.insert(student1);
 				System.out.println(insertResult + " Row inserted");
@@ -77,22 +78,22 @@ public class App {
 
 				System.out.println("Enter New City:");
 				String city2 = sc.nextLine();
-				
+
 				if (id2 <= 0) {
-				    System.out.println("Invalid Id. Must be greater than 0.");
-				    break;
+					System.out.println("Invalid Id. Must be greater than 0.");
+					break;
 				}
 
 				if (name2.trim().isEmpty()) {
-				    System.out.println("Name cannot be empty.");
-				    break;
+					System.out.println("Name cannot be empty.");
+					break;
 				}
 
 				if (city2.trim().isEmpty()) {
-				    System.out.println("City cannot be empty.");
-				    break;
+					System.out.println("City cannot be empty.");
+					break;
 				}
-				
+
 				Student student2 = new Student(id2, name2, city2);
 				int updateResult = dao.change(student2);
 				System.out.println(updateResult + " Row updated");
@@ -101,18 +102,31 @@ public class App {
 			case 3:
 				System.out.println("Enter Id to delete:");
 				int deleteId = sc.nextInt();
+				
+				if(deleteId<=0)
+				{
+					System.out.println("Invalid Id. Must be greater than 0.");
+					break;
+				}
 
 				int deleteResult = dao.delete(deleteId);
 				System.out.println(deleteResult + " Row deleted");
 				break;
 
 			case 4:
-				
-				System.out.println("Enter Id:");
-				int id = sc.nextInt();
 
-				Student student = dao.getStudent(id);
-				System.out.println(student);
+				System.out.println("Enter Id : ");
+				int id = sc.nextInt();
+				
+				if(id<=0)
+				{
+					System.out.println("Invalid Id. Must be greater than 0.");
+					break;
+				}
+				
+				Student singStudent = dao.getSingleStudent(id);
+				System.out.println(singStudent);
+				
 				/*
 				 * List<Student> students = dao.getAllStudents();
 				 * 
@@ -122,6 +136,16 @@ public class App {
 				break;
 
 			case 5:
+				List<Student> students = dao.getAllStudent();
+				System.out.println("--------------");
+				System.out.println(students);
+				System.out.println("--------------");
+				for (Student s: students)
+				{
+					System.out.println(s);
+				}
+				break;
+			case 6:
 				System.out.println("Exiting program...");
 				sc.close();
 				System.exit(0);
